+++ app/src/api/index.ts
@@ -1,7 +1,6 @@
-import axios, {AxiosError, AxiosRequestConfig, AxiosResponse} from 'axios';
+import axios from 'axios';
 import type {
     ApiAddToLeaderboardRequest,
-    ApiBadRequestError,
     ApiChangePasswordRequest,
     ApiGetLeaderboardRequest,
     ApiGetLeaderboardResponse,
@@ -16,9 +15,7 @@ const API_URL = new URL('https://ya-praktikum.tech/api/v2');
 /**
  * Клиент для API https://ya-praktikum.tech/api/v2/swagger/
  */
-const ax = axios.create({baseURL: API_URL.href, withCredentials: true});
-ax.interceptors.request.use(requestHandler);
-ax.interceptors.response.use(responseHandler, responseErrorHandler);
+export const ax = axios.create({baseURL: API_URL.href, withCredentials: true});
 
 /**
  * Запрос регистрации
@@ -120,60 +117,6 @@ export async function getLeaderboard(data: ApiGetLeaderboardRequest): Promise<Ap
     const response = await ax.post<ApiGetLeaderboardResponse>('/leaderboard/all', data);
 
     return response.data;
-}
-
-/**
- * Перехватчик запросов на сервер. Позволяет выполнять обработку всех запросов
- * до отправки в сетевой стэк
- * @param request   Объект конфигурации запроса на сервер
- */
-function requestHandler(request: AxiosRequestConfig): AxiosRequestConfig {
-    // eslint-disable-next-line no-console
-    console.log('[API req]', request);
-
-    return request;
-}
-
-/**
- * Перехватчик ответов от сервера. Позволяет выполнять обработку всех ответов
- * до возвращения в промис вызвавшего метода
- * @param response  Объект ответ от сервера
- */
-function responseHandler<T>(response: AxiosResponse<T>): AxiosResponse<T> {
-    // eslint-disable-next-line no-console
-    console.log('[API resp]', response);
-
-    return response;
-}
-
-/**
- * Перехватчик ошибок от сервера. Позволяет выполнять обработку всех ошибок,
- * до возвращения в промис вызвавшего метода
- * @param response  Объект ошибки от сервера
- */
-function responseErrorHandler<T extends unknown>(responseError: AxiosError<T>): AxiosError<T> {
-    let badRequestError: ApiBadRequestError;
-
-    if (responseError.request.responseType === 'text' && typeof responseError.response.data === 'string') {
-        try {
-            badRequestError = JSON.parse(responseError.response.data);
-        } catch (error) {
-            // Несмотря на описание Api, формат ошибок иногда не такой как должен быть
-            badRequestError = {
-                reason: responseError.response.data
-            };
-        }
-    } else {
-        badRequestError = responseError.response.data as ApiBadRequestError;
-    }
-
-    if (badRequestError) {
-        badRequestError.status = responseError.response.status;
-    }
-    // eslint-disable-next-line no-console
-    console.error('[API error]', badRequestError || responseError);
-
-    return responseError;
 }
 
 /**
